#35 Corregido bug relacionado con la progress bar de Series.

+++ src/main/java/com/furyviewer/repository/FavouriteSeriesRepository.java
@@ -33,7 +33,7 @@ public interface FavouriteSeriesRepository extends JpaRepository<FavouriteSeries
     Boolean selectFavouriteSeries(@Param("id") Long id);
 
     @Query("select count(f.liked) from FavouriteSeries f where f.liked=true and f.series.id=:id")
-    Long countLikedSeries(@Param("id") Long id);
+    Integer countLikedSeries(@Param("id") Long id);
 
     @Query("select s.liked from FavouriteSeries s where s.series.id=:id and s.user.login = :login")
     Boolean selectFavouriteSeriesAndUser(@Param("id") Long id, @Param("login") String login);
